+++ lib/faker/internet.ex
@@ -128,4 +128,20 @@ defmodule Faker.Internet do
       |> String.rjust(2, ?0)
     end) |> String.downcase
   end
+
+  @doc """
+  Generates a slug
+  If no words are provided it will generate 2 to 5 random words
+  If no glue is provied it will use one of -, _ or .
+  """
+  @spec slug([String.t], [String.t]) :: String.t
+  def slug(words \\ nil, glue \\ nil) do
+    words = words || Faker.Lorem.words(2..5)
+    glue  = glue || ["-", "_", "."]
+
+    words
+    |> Enum.take_random(length(words))
+    |> Enum.join(Enum.random(glue))
+    |> String.downcase
+  end
 end